Ajustes para exibir o valor de venda.

+++ src/main/java/br/com/sebastiao/junior/ecommerce/dto/ValoresAdicionais.java
@@ -11,5 +11,5 @@ public class ValoresAdicionais {
 
 	private Double despesasTotais;
 	
-	private Integer margemLucro;
+	private Double margemLucro;
 }
